+++ Mono.Cecil/AssemblyWriter.cs
@@ -1487,7 +1487,7 @@ namespace Mono.Cecil {
 			if (type.HasNestedTypes)
 				AddNestedTypes (type);
 
-			if (symbol_writer != null && type.custom_infos != null)
+			if (symbol_writer != null && type.HasCustomDebugInformations)
 				symbol_writer.Write (type);
 
 			WindowsRuntimeProjections.ApplyProjection (type, treatment);
